feat: add multiple relay URLs in NostrFetcherTrait
This commit introduces support for multiple relay URLs in the NostrFetcherTrait. The application will now attempt to fetch data from a list of relay URLs instead of just one. This change provides a better failover mechanism in case one of the relay servers is down or unresponsive. The application will now throw a RuntimeException if no data is received from any of the relays.